Added bonds and angles as data types
Modified the Datum class to include str_num. Keeps track of the
structure/conformer number for large mae or mmo files. Modified
the sorting methods accordingly.

The previous change was integrated to accomodate the new data
types: bonds and angles. These data types work on files containg
a single structure or many structures. The str_num attribute was
added to organize this data. Previously only energies were fully
compatible with multiple structures per file. Now, to the best of
my current knowledge and testing, everything works with multiple
structures per file. For this to work, many adjustments were made
to the calculate script.

These new types are accessed via "-ma" and "-mb" for MacroModel
angles and bonds, and "-ja" and "-jb" for Jaguar angles and bonds.
Unfortunately, these data types are extracted from an mmo file, so
a calculation must be performed even for the reference Jaguar data.
Extraction from the mae file is possible, but would require a lot
more work. Now that I think about it, it also probably couldn't
fulfill my point in the next paragraph.

Currently, all bonds and angles are used. However, this should
be limited to bonds and angles that are actively being
parameterized. All other bonds and angles should not be included
the data. This change will come soon.

Added the option "--norel" to calculate. This prevents converting
absolute energies to relative energies for printing/writing data.
Relative energies are still used no matter what when calculating
the penalty function (that just wouldn't make sense otherwise).

Minor changes were made to the output from calculate to
accomodate the now longer data labels.

The calculation indices determined when creating com files in
calculate.make_macromodel_coms are now split in two: one for mae
files and one for mmo files.

The ELST, WRIT, and MINI commands still may be redundant. However,
things seem to work now, and I'm tired, so it is what it is for
now. Perhaps later someone will optimize this further (if
possible).

Some changes were made to filetypes.BaseFile.get_data to be more
accomodating to the multiple structure format that is now assumed.

Minor additions were made to filetypes.MMoFile to finish the data
extraction.