Added MM3* Hess. Added Jaguar Hess. mass-weighting
Added options to calculate.process_args and the rest of the script
to handle getting the MM3* Hessian. Also modified the Jaguar Hessian
method in calculate.extract_data to account for mass-weighting.

Modified the Hessian methods of filetypes.FileType to make them more
convenient.

The current method appears to work, but it feels sloppy. I think that
we should make a class for the Hessian that does a better job at
converting between mass-weighted versions, inversion, etc.