Modified bond/angle extraction
Bonds and angles are now only included as reference data if there
is a FF term that directly relates to them. However, this distinction
requires the mmo file. Each line pertaining to a bond or angle in the
mmo file has a comment at the end. For substructures in MM3*, that
comment refers to the substructure's name.

Right now, all bonds and angles are eliminated based on the keyword
"OPT", which must appear the in the substructure's name. It would be
nice to auto-detect the substructure name.