Minor change to simplex.
If the simplex optimizer is working with very similar paramter
sets (perhaps the simplex has contracted many times and the
parameter sets are numerically equivalent) or even different
parameter sets that happen to all produce the same objective
function value, then in the weighted inversion scheme, a
ZeroDivisionError will be raised. Added error handling for
this.