Fixed bug where simplex wouldn't update parameters.
Simplex is supposed to return the best FF obtained.
If the number of parameters to be optimized is
greater than the argument max_params, then how the
final parameters are returned is different. Fixed
a bug where parameters wouldn't actually be
updated when this method is used (when params >
max_params).