Fix for issue #2.
Each MM3* parameter has what I call a mm3_label. It is the 1st 2
characters of the line containing the values for a given parameter. In
your case, "a2" is the mm3_label. The "a" comes from geometric
dependence on when to use that parameter.

datatypes.py contains the MM3 force field class, and it has the methods
import_ff and export_ff, which do exactly what their names suggest. The
import_ff method worked with these geometry dependent parameters, but
export_ff failed to recognize them. I have since updated datatypes.py
to let it handle this sort of issue. I also added a short test script
to check parameter labels (this is my first of hopefully many tests
that will be implemented).

The MM3 class now has a number of methods that check mm3_label using
regex. I'm still unsure whether these methods are best located here,
but this at least works for now.