Working on reading Gaussian archive.
So far, I can efficiently read, well, lots of things. As for the
Hessian, I'm not sure if Gaussian is giving me the lower or
upper triangular form. I'm also not sure of the units. And I'm
also not totally sure what to do with the eigenvalues in the
archive. At least, I think those are the eigenvalues.

Still a work in progress.