Added boolean on/off switch for selecting which charges to use.
Now, in the .mae files, you may include an atom property,
"b_q_use_charge". This is a boolean (0 or 1) property. Here's how
it works.

You may use -mq or -jq just like before with the calculate module.
If an atom doesn't have the property "b_q_use_charge" or if the
property "b_q_use_charge" is 1, then the partial charge of that
atom will be considered as a data point. If "b_q_use_charge" is 0,
then that charge is ignored.

This is progress towards issue #7, which raised concerns as to which
partial charges should be included in the optimization. It's now
up to the user to make that decision, and it's very flexible.
The drawback of this method is that it requires the user to edit
their Maestro files (.mae), which may be fairly time consuming.

An alternative method would be to use a separate text file
containing the atom numbers for the atoms whose partial charge
we want to include. I may add this as an option to the calculate
module later.

Another method would be to automatically detect which atoms have
custom dipole parameters by generating a MacroModel (.mmo) file,
but that would require an additional MacroModel calculation in
the case of -jq (if -jq was used alone), and may slow down -mq.