Removed force field assertion.
Certain data types (structural data from .mmo files comes to mind)
require knowing the name of substructures or parameter comments.
To get these names,the user must either provide them, or the code
must read the force field file.

Previously, if the comments/substructure names weren't provided,
it would assert that the user gave the path to the force field.
Instead, it now checks for mm3.fld in whatever directory was
provided. If no directory was provided, it checks the current
working directory.