Changed how ParamMM3 returns reference angles.
Previously, angles could adopt values above 180 degrees. This
naturally occurs as the optimization techniques are non-constrained.
Constraining parameters to positive values occurs after the fact,
i.e. they are assigned a negative value, and then an exception is
raised indicating that this shouldn't be allowed. It is then dealt
with.

Now, angles do a similar check if the reference angle is set higher
than 180. The value is still allowed to go to, say, 183, but whenever
the value is returned, it will be returned as

    returned value = value - abs(180 - value).

In the above example (183), the returned value would be 177.

This probably isn't as ideal as constrained optimizations, but
hopefully it works.