It appears that this is due to some code in pHessMan.py.
Options -geigz and -mgeig call upon the filetypes.GaussLog.evals or
filetypes.GaussLog.evecs, respectively. Calling upon either of these
calls filetypes.GaussLog.read_out(). This then calls upon
pHessMan.readgautout(self.path).

This part actually functions as intended.

The problem is that the argument parsing and the code that is executed
when using the pHessMan module from the command line is not contained
in the __main__ namespace, so it is executed immediately upon loading
the module, which calculate does. When loading that module, pHessMan
also process the arguments for the calculate module, and throws the
"Unrecognized option:" warning.

Seeing as the only part of pHessMan that I need is the readgauout
function, I'm just going to move that into the GaussLog class as its
own function. I have been meaning to do this for awhile anyway.

Alright, it's now inside filetypes.GaussLog.read_out(). The syntax is
still confusing, more or less a direct copy of what was in pHessMan.
Anyway, it works.

Closes #12.