Many changes to hopefully improve the memory usage of gradient.
This is still untested and unrefined code.

Previously, during central differentiation of parameters, all of
the data points for an incremented or decremented parameter FF
were stored in memory.

Now, they are stored in a text file. This file is then read for
data two lines at a time (the incremented and decremented data
points for a single parameter).

Currently, the format of this parameter file is:
1 - Labels
2 - Weights
3 - Reference data points
4 - Original FF data points
5 - Parameter 1 incremented data points
6 - Parameter 1 decremented data points
7 - Parameter 2 incremented data points
8 - Parameter 2 decremented data points
...