Modified generation of .com files.
The calculate module now accepts the argument "--append" or "-a".

For example,

    python calculate.py -d /somedir -mb somefile.mae

generates the file "/somedir/somefile.q2mm.com." Now you can
use the command

  python calculate.py -d /somedir -mb somefile.mae -a dragon

to instead generate a command file with the name
"/somedir/somefile.dragon.q2mm.com."

Also, command files will not be overwritten if they already exist.
This gives the user a chance to modify the command file (such
as adding in solvation treatement).