Modified behavior of writing MacroModel .com files.
Previously, in commit 969894ef24349d48c2, I made it such that
a q2mm.com file will never be overwritten. This allowed for
modification of the .com files to include things such as solvation
models.

However, it also necessitated that old .q2mm.com files be deleted,
and that the reference and calculated commands each used the
"--append" or "-a" command to ensure they generated differently named
.com files. For example,

  python compare.py -c " -mb X001.mae -a cats"
                    -r " -jb X001.mae -a dogs"

would work properly. However,

  python compare.py -c " -mb X001.mae" -r " -jb X001.mae"

would not work. The calculate module would generate a .com file for
the option "-jb", but this would not be overwritten with the .com
file for the "-mb" option.