Working on script to setup ESP calculations in Jaguar.
Added a new piece of code, setup_esp.py. Still debating whether or
not this should be its own module.

Anyway, his takes a collection of files (Jaguar .in for a template,
MacroModel .mae for MM charges, and a MacroModel .mmo for determining
what atoms are involved in bonds that are being optimized). The script
adds an atomic section to the Jaguar .in file.

This new atomic section has the partial charges for every atom that
MacroModel is not optimizing dipoles for (this is determined using
the substructure name in the MacroModel mm3.fld file).

This .in can then be used (with some modification of the &gen section)
to run a Jaguar ESP optimization, where all partial charges that we
can't optimize through MM3* dipoles are set to whatever the current
MM value is.