Updated Schrodinger token check.
We look for specific strings using regex from the output of
Schrodinger's licutil to identify whether there are enough
available tokens to run a job.

For at least some versions of Schrodinger, when no tokens are in
use, Schrodinger will not report anything (rather than saying 0 out
of x tokens in use, it just says nothing).

Modified the license check function to tolerate this behavior.