Working on removing sorting and improved import of reference data.
Mostly made changes to calculate. Data types will always be calculated
in the same order now. This is to ensure that reference and FF data
sets can be aligned without needing sorting.

Sorting is being removed to:
 1. Make using a reference file easier.
 2. Speed up code.

Currently, only implemented -me and -je. These are actually the most
complicated data sets to incorporate since the FF data needs to be
adjusted relative to the reference data. Hence it's always a good idea
to start with this and make sure this works.

From here on out, it will be simple to implement the remaining data
types.

Also removed some functions from compare that are now unnecessary.
For example, for the average energy approach, that was previously done
in compare. That will now be handled in calculate. Not having to look
up the data a second time in compare makes everything more efficient
anyway.