Added all MacroModel and Jaguar energetic datatypes to calculate.
Within a given backend, all energies can be used together. i.e. -je,
-jea, -jeo, -jeao can be used all at the same time. The same goes for
the MacroModel equivalents. However, don't mix Jaguar and MacroModel.
I'm not sure why you would ever do that, but I haven't tested what
happens when you do yet.

Adding Gaussian next.

There is currently a ton of repetition inside calculate. In the
future, it would be wise to add some functions to avoid all the
repetition. For now, I'm just leaving it as is because it's very
readable and simple for me to implement.