Added back in Jaguar and MacroModel structural data.
Right now, I'm just addressing the sorting issue. The code herein
is more or less a direct copy of the previous version, but modified
to work without the sorting paradigm.

Expanded the regex for substructure name matching in the constants
module to account for scenarios where the relatively untested
torsional check would fail.

Unforunately, there is very little documentation on the torsional
check.

How the torsional check works:
 1. Figure out atom numbers in a torsion.
 2. Break that down into two angles.
 3. Check those angles in the given structure to make sure neither is
    near 180 degrees.
 4. If either is near 180 degrees, exclude this torsion.

The torsional value for a torsion that has an inside angle near 180
degrees can fluctuate wildly. We don't want to include such sensitive,
unstable data in the optimization.

I would like what's added in this commit to be cleaned up a bit in
the future. I think that filetypes classes could use some changes as
well. The whole select_structures and organization of structures is a
bit complicated.

Also, the torsional check functions need documentation and cleaning
up.

WARNING: I didn't double check for any errors in the recent changes
         to torsions! Always double check your data before starting
         an optimization, and watch out for the torsions data type!