Added in Jaguar and MacroModel charge data.
Options -mq and -mqh function very similar to how they did previously.
The only difference is that all of the charge options, including the
new -mqa (read below), can be used simultaneously.

Also added option -mqa. This removes all hydrogens that are connected
to one atom from the data. Also, it sums the charge of that hydrogen
into the atom it's connected to.