Clean up, and feature addition to loop.py
Sorting was still present in RDAT and CDAT, and so those lines of code were removed.

Added section with COMP command now allows you to just print out results on the front end if you do not put any output option.

Added a new command for a loop input file called WGHT or weight. This allows as user to adjust the default weights of a data type without having to edit the constants.py textfile.