Restart gradient optimizations from central differentiation file.
Added in an option to restart gradient optimizations from the central
differentiation file.

Currently this doesn't work with NR because the parameter derivatives
are stored on the central differentiation FFs, which are never generated.

Still need to test this more and add in functions such that works
with the loop module.