Improved Gaussian archive reading regex.
Gaussian archive starts with " 1\1\". It ends with some number of \
followed by @. Not sure how many \ there are, so this matches as many
as possible. Also, this could get separated by a line break (which
would also include adding in a space since that's how Gaussian starts
new lines in the archive).

Previously, we didn't account for the line break occuring between the
final \ and @.

Fixes issue #10. In Neil's problematic file, the last line of the
archive was simply " @".