Modifications to allow multiple energies in Gaussian log files
Untested, but these are probably all the changes needed in filetypes.
This was a little trickier than I expected because using z-matrix
coordinates adds another section (separated by \\) to the archive.

Also, I didn't write any code to read z-matrix coordinates. The atoms
are simply skipped for now.

Now the problem is that we have to account for having HF vs. HF +
ZeroPoint energies.

Changes to address issue #12.