Corrected the com property for several data types
I initially wrote and tested the modifications for the command ge,
which seem to be working well.

When I extended this to gea, geo and geao, I more or less copy and
pasted the code from ge, and forgot to update the com property of
the Datum instances.

Corrected in this commit. Also, this is to address issue #12.