Correcting behavior of SVD
See the discussion on issue #28 for more details. I will try to
recap some because this discussion will unfortunately not transfer
to forked repositories of Q2MM.

SVD should do the following.

    J = U . s . V
    x = VT . s' . (UT . r)

Captial T stands for transpose. s' is a modified version of s. J is
the Jacobian and r is the residual vector.

As described by Per-Ola, "The beauty of it is that, since you don't
invert the small values, but set the inverse equal to zero
(counter-intuitive), you get rid of linear dependencies between
parameters, and all parameters with only small influence on the
objective function."

With the modification in this commit, I use np.linalg.svd where the
argument full_matrices is set to False. This is the only way I could
get the matrix multiplication to work with the above formula.

However, there is commented out code for using the full matrices,
which of course fails when you attempt to solve for the parameter
changes, x.

Previously, the code did this.

    A = JT . J
    b = JT . r
    A = U . s . V
    A' = U . s' . V
    A' . x = b