Added options to invert Hessian
The Hessian can now be inverted using the command -i somefloat for
calculate. This applies to -jh, -gh, -jeigz, -geigz.

For -jh and -gh, the Hessian is decomposed into its eigenvectors and
eigenvalues using numpy.linalg.eigh, the eigenvalue is inverted, and
the Hessian is reformed.

For -jeigz, we read the Hessian and eigenvectors. This is used to
form the eigenvalue matrix. The diagonal is extracted (all off
diagonal elements should be nearly zero), the lowest value is
inverted, and then the eigenvalue matrix is reformed as a strictly
diagonal matrix.

For -geigz, we simply read the eigenvalues. The lowest value is
inverted, and the eigenvalues are used to form a diagonal matrix.

Addresses issue #30.