Fixed variables being used before declared
Each MM3* parameter has the attributes ff_row and ff_col. If the number of
parameters has to be reduced during a simplex optimization, simplex will save
the values of the MM3* FF row and column for the parameters with the lowest 2nd
derivatives. These are then used later on to reduce the number of parameters in
various FFs.

Note that this method is specific to MM3*, and should be replaced with a more
general method that will work with other FFs.

Anyway, the lists that contained the FF rows and columns, ff_rows and ff_cols,
were declared within an if statement, and sometimes they have to be used outside
of that if statement. This commit moves their declaration outside the if
statement.

Also added more comments.

Working on issue #22.