Steps and inversion point
Too few steps taken now that the Simplex is working. In particular,
the number of steps w/o change now depends on number of parameters.

The division by zero in determination of weighted inversion point
did only break the for loop, not the while loop. Instead, moved
the weighting calculation outside the foor loop (also more economical),
and used a zero test instead of a try: statement.