Maximum cycles w/o change no longer set by user
Changes to make the maximum number of cycles without change equal to 3 * N,
where N is the number of parameters in the simplex. The maximum number of
cycles can no longer be set by the user and the variable appears in
`simplex.__init__`, so it is now an internal variable. I adjusted the name to have
a leading underscore as such.

In the future, it would be wise to change `self._max_cycles_wo_change` back
to `self.max_cycles_wo_change`, and then also add something like this

```
if not self.max_cycles_wo_change:
    self.max_cycles_wo_change = 3 * (len(self.new_ffs) - 1)
```

This would give the user the option to override the default if he or she desires.