Correcting behavior to exit simplex
It's possible that simplex will generate a set of trial FFs where
all of the scores are the same. At this point, simplex should exit.

Previously, there was a break instead of a raise that prevented
simplex from exiting properly.

Also added in another layer of checking thanks to Per-Ola's
suggestions. Now there are two chances to catch the potential
error.

Working on issue #22.