Modified behavior of datatypes.MM3.lines
There are FF objects, instances of datatypes.MM3, that have the attribute lines.
datatypes.MM3.lines is a list of strings, literally every line from the mm3.fld
file. Since this is rather large, I try not to store it in memory more than
once.

By the way, datatypes.MM3.lines is used to write FFs in datatypes.MM3.export_ff.

datatypes.MM3.export_ff accepts several arguments. The following is copied from
its __doc__.

    def export_ff(self, path=None, params=None, lines=None):
        """
        Exports the force field to a file, typically mm3.fld.

        Parameters
        ----------
        path : string
               File to be written or overwritten.
        params : list of `datatypes.Param` (or subclass)
        lines : list of strings
                This is what is generated when you read mm3.fld using
                readlines().
        """

The code used to check if lines and self.lines was None, however, self.lines is
now actually a property. When self.lines was updated to being a property, this
negatively affected that behavior.

For example, if you were to say

    if self.lines is None:
        do_stuff()

This would actually call upon the property self.lines. This property looks like
this

    @property
    def lines(self):
        if self._lines is None:
            with open(self.path, 'r') as f:
                self._lines = f.readlines()
        return self._lines
    @lines.setter
    def lines(self, x):
        self._lines = x

This means that you can call upon self.lines, it will check if self._lines is
None. If it is, then it attempts to read the lines. This would cause problems
when self.path was also None because it couldn't open the file to read the
lines.

Basically, self.lines was being called when self.path was None, which resulted
in errors.

This works on issue #22. There's also a very good chance this will fix many of
the problems Per-Ola has been reporting. More fine tuning may be needed, but I
think this is the core issue.