Changed how weights are imported
Previously, each data point would be checked individually for whether
or not the weight was set. Now, it just checks the 1st data point. If
the weight is present on the 1st data point, it assumes the weights
are also set on all the other data points.

This was done to speed up the code.