Improved behavior of the loop cycles
If the COMP command wasn't supplied before LOOP, then the code would die because
there would be no existing FF score.

Before looping begins, I added a check to ensure that we know the existing FF
score. If it doesn't exist, it throws a warning and calculates the score itself.

Addresses issue #34.