Fixed issue where FF backups were being written multiple times
FF backups are made at the end of each loop cycle. Previously, this would make
the backup in the directory specified by the user ("DIR" from the loop input
file) as well as the current working directory. Now, it only makes the backup in
one directory. This is the one specified by the user or, if none is specified,
the current working directory.