Working on problem with projected keyword in Gaussian
Using the projected keyword in Gaussian seems to add in an additional line that
includes the key phrase "IRC Coupling". This was throwing off the reading of the
eigenvectors.

I believe that adding the simple checks for that key phrase and skipping that
line if present seems to have fixed the issue.

Also commented out a lot of logging.

Still could use more testing.

Reference issue #19.