Fixed issue when initial simplex score was the same as final
When the initial simplex score was the exact same as the final, simplex would
say

    best_ff = self.new_ffs[0]
    ...
    return best_ff.

That's the same FF as the original FF, so returning best_ff was okay. However,
it wouldn't update best_ff to contain the full parameter list.

Now, if there's no improvement in the scores, then it will do

     best_ff = self.ff
     ...
     return best_ff

This ensures that the full parameter list is retained.

Fixes issue #27.