Modified simplex parameter selection
Defined a simplex variable:
    (2nd der.) / (1st der.)**2.

Parameters are sorted according to this variable. The lowest parameters are
selected for simplex optimization.

Parameters with values larger than 1 shouldn't be used in simplex. These
parameters are bracketed by the differentiation and should be left to the
gradient optimization. Accordingly, I wrote some code for that in this commit,
but it's currently commented.

Eliminating all parameters with simplex variables higher than 1 could lead to
exceedingly small parameter sets, perhaps 1 or 2 parameters.