Fixes bug related to score calculation and ref file
When a file is used for the reference data of a torsions--instead of calculating the reference data--there would be a "NONE" attribute to r_datum.typ. This resulted in penalty scores that failed the "if r_datum.typ == 't':" logic. By changing r_datum.typ to c_datum.typ in this commit change we avoid failing the logic and continue to calculate the score as intended.