Added tool for merging structures
Takes a set of input structures and merges onto them a single other structure.

The input structures must all have the same numbering for this to work nicely.
In the future, I'd recommend adding identifiers inside the .mae files to manage
using totally different input structures.

Only one structure can be merged onto the input structures. This must be
contained in a separate file. In the future, it should be easy to expand this
into making a matrix of combined structures. I didn't write it because I just
wanted to get done what I had to as quickly as possible. I'd recommend making
the adjustments mentioned above (use more information in the .mae) before
making this change.

Contains options to delete atoms from the input structures and the merged
structure. Typically, these would be any atoms that overlap.

Can form bonds based upon the user's input between the input structures and the
merged structure.