Initial interfacing for tinker
By using the formats of for macromodel initial interfacing for tinker has been setup.

Calculate.py: This includes adding the commands for argument parser and basic set up of commands associated with tinker files to determine what is calculated.

Filetypes.py: Added the Tinker_xyz class that has functions to run tinker calculations. So far this code works for single point calculations. As with macromodel we will have to allow the code to combine all commands for a file to be combined in one calculation, or atleast outputed to the same output file. From what I can tell from tinker, there is little control or ability to print to an output file therefore I am using sp.checkout(args, shell=True) to run the calculation and print or append to an output file. It would be better to use tinker's output control if it exists that way we won't have to use "shell=True"