vs.py seems to be working for intended purposes
Still needs to be formatted to accept combinations of ligands. Basically, need a
base template, but this should be able to handle adding on an arbitrary number of
additional structures.

The inversion from r to s chirality could be simplified. Probably could reduce
the amount of code there fairly easily.

Need to encapsulate the main function.