ParamError can be raised when loading parameter files
Force fields can be loaded in a number of ways. Ranges for parameters are
loaded after the parameter values have been loaded. If the parameter value
is outside the range when the range is loaded, this wasn't caught until much
later on.

Modified the behavior to check for valid parameter values immediately upon
loading parameter ranges.