Updated rhodium hydrogenation of enamides FF
Added another version of the rhodium hydrogenation of enamides FF. The
parameters are the same as Donoghue's. However, I reformatted the substructure.

Specific atom types for the hydrogens and phosphorus atoms are no longer
requred.

This was done using the geometric parameter selection provided by MacroModel.
See the MacroModel documentation for more details.