Removed custom settings for MCNV
See comments on the changes in the code below.

What if we just leave this out? Sets range for allowed DOF to change.
Here, it says change 1 - 5 DOF. This seems low. Default is 1 - N,
N is the number of variable dihedral angles. At any rate, these
initial values are overridden by the "adaptive mechanism" as the
search progresses. Good luck finding any explanation of what the
"adaptive mechanism" is from any of the Schrodinger documentation.
The "adaptive mechanism" can be turned off with DEBG 103.