Tinker calculate, collect, and compare data
First the setup that I had for the hessian calculation was wrong. I was
performing the calculation on filename.q2mm.xyz which in everycase would
have been a TINKER optimized structure. So instead I changed the variable
name to use filename.xyz which would be the QM reference structure in
TINKER format. This also introduced problems when it came to writing a
.hes file. TINKER will always write filename.hes for filename.xyz regardless
if an output file is give (i.e. filename.q2mm.hes). I added logic to fix
this problem without rewriting a filename.hes if it does exist. All q2mm
generated files should be named with .q2mm in the filename.

All data, except for charges (which we should't need), can be calculated,
collected and compared. The hessian in TINKER is not mass-weighted and
has kcal as the energy unit instead of kJ. This is now fixed. Additionaly
I have fixed the gaussian hession data to be mass-weighted. As for energies,
I'm pretty sure they are in kcal in TINKER but I'm unsure if this energy is
converted to kJ in the Q2MM code. I'll look into this later.

But it should also be noted that I haven't yet fixed the mixed up ordering
of geometeric data (This is described in the previous commit). Johnathan
will be helping out with this.