Corrected Geometric Ordering Problem
The problem that is mentioned in the previous two commits has now
been fixed. As the geometric data is being gathered from the
tinker.q2mm.log file, for each structure the data type will be sorted
based off of the way we set up our mm3 force fields in MacroModel.
So bonds are sorted by the first atom (which should be the lowest of the
two). Angles are sorted based on the middle atom and then the first atom.
Torsions are sorted based off of the second atom and then the third,
first, and then forth. So far this has proven to be successful.