Added logging and Comments
I added in logging to match similar to what is used in other areas
of the current code along with adding comment lines in each of the
tinker added functions.

Additionally I removed the charge arguments (tq, tqa, tqh) b/c
TINKER does not calculate charges. I have also changed the output
of the potential energy that TINKER provides to be in kJ/mol instead
of kcal/mol.