Fix bug with removing other matches
Without going into the details of why this is happening, I have noticed several instances where first_match_only would be true but several matches would be returned. 
Here are examples of the first_match_only property results from the schrodinger package.

If there is a pi-bound ring to a metal center that is described by a dummy atom the substructure that we use is along these lines:
M-D1.CR-CR(.2)-CR(.2)-CR(.2)-CR(.2)(.3)
Due to the symmetry there would be a total of 5 matches. With the first_match_only property schrodinger will just return one.

For a system where the first_match_only property doesn't work. This is a case where the N3 atom has two H3 atoms bound to it:
H3-N3.M.N0
The two H3 atoms are different, in that one is an innocent H3 and the other is involved in a transfer hydrogenation TS. The schrodinger property does not seem to eliminate one of the two matches, which is why we need to implement this additional logic.