Atom listings for M-Cp and M-Ph
This probably isn't the best logic, but it should be quick enough that it doesn't matter. In instances where there are dummy atoms that are bound to several atoms (6+), the metal and and dummy atoms need to appear at the bottom of the atom listing in order for any minimization or conformational search can be performed. This code reorders the atom listings to accomidate for this weird bug in macromodel. There are a few problems that may arise from this:
1. The atom iterator is preformed regardless, and the code isn't very general (i.e. I have included specific atom_type_names.)
2. Perhaps a user would like to keep certain atoms in a certain listing, and this would change that regardless of the structures present.