Quick fix for updating RCA4 and TORC after reorder
After any renumbering of any of the atoms, the RCA4 and TORC values need to be corrected to the new atom index value. These are a quick fix, that just gets the job done. This problem goes back to using numbers instead of atom objects for these bond properties, which is found in issue 42.