Prevents unrealistic merge of aromatics and merge update
Some instances an aromatic group will be matched based off of a
relativly inclusive matching pattern. Two examples are for the
Heck reaction and the osmium dihydroxylation reaction. When
matching the alkene with the substrucutre (C2=C2-00) an aromatic
group would be matched which would give (in the case of benzene)
3 matches for double bonds, with four different matches per bond
giving a total of 12 additional structures. This code iterates
through the rings and matches to determine those matches that
are within the ring (2+ atoms in both matches and the ring) and
then determines if it is aromatic or not.

Also the code would prevent structures that are intended to have
multiple matches such as (2,1,5) and (5,1,2) to use both matches.
To fix this I added logic to determine if the 'b_cs_first_match_only'
is true before determinig if the matches are unique or not.