Fixed bugs from previous commit (#39)
* Removed print lines

Forgot to remove excess print statements

* Fixed bugs from previous commits

I had to add in another structure property to differentiate between
aromatics we want, and those that we don't want. This propert is
called "b_cs_accept_aromatic". Additionally, there was a problem
where a match would be duplicated because two of the atoms in the
match belonged to two rings, or something along this line. So now
there is logic to determine if the match needs to be appended or
not.