Added more mini and mcmm after merging
Noticing several structures that were still not realistic structures,
I added more minimizations and conformational searches without forzen
atoms at the end of the merging sequence. The problem is that these
now take forever since everystucture will have conformatioanl searches.
I think a better way would be to base it off of the convergence of the
minimization or as groups are added onto the previous structure then
reduce the number of conformational steps needed? I'm not sure this is
too much of a problem.