Add Ligands, Substrates, Nucleophile, and Template for Heck.
First I reorganized the substrates and ligands directory to
contain subdirectories based off of ligand type or substrate
type. I then added in two new folders in the substrates: 1)
alkenes and 2) aryl_alkenyl. The alkene folder contains
a variety of olfins and the aryl_alkenyl folder contains
nucleophiles that are bound to a palladium center. Within the
ligands folder I added in a new ligand class called pyrox for
pyridine-oxazoline. Addtionally I added two templates for the
Heck reaction which are differentiated in a readme file in that
folder. I will add ChemDraws into each of the heck-related
folders that show the structure and filename.