Added and updated parts for VS.
I did not add a few alkenes to the folders in the previous commits and
there were some ring closure errors associated with some of the alkenyls.
I'm not sure what the problem with the ring closures was, but in general
we should be cautious when it comes where ring breaks are located with
respect to other ring breaks in the system.

Additionally I am adding a new tool that will create gaussian command files
from *.mae or *.maegz. This is new and I am still working on it, but this
will be completed soon.