Added logic for Frozen atoms and the calculation type.
Frozen atoms are based off of internal coordinates and are a single
string with ';' seperating different patterns. The patterns are then
used with the schrodinger structure utility that we have used in the
past. I opted to remove the dummy atom, if present before this step
so that the atom numbering is correct in the gaussian *.com file.
I'm not sure if this will be stable for all cases we are interested in.

Additionally I have set up the logic to write the route section of the
*.com file. In the case of a calcualtion with frozen coordinates a
seperate link1 section will follow the first calculation that does
not freeze any coordinates. I also want to include a option to
have gaussian write a *.wfn file. This file is the wavefunction and
can be used with the NCIPlot program to give some pretty interaction
figures.